scx_lavd: Refactor out FlatTopology from main.rs.
As a first step toward energy model-aware CPU ordering, let's refactor
FlatTopology by separating it into a new module, cpu_order.rs. As the name
'FlatTopology' no longer makes sense, let's rename it to 'CpuOrder'.
Additionally, clean up struct names for consistency (e.g., CpuFlatId ->
CpuId, etc.).

There are no functional changes.
